requirements-test.txt
astroid==1.3.6
cov-core==1.15.0
coverage==3.7.1
flake8==2.4.1
IMPLib2==0.9.2
logilab-common==0.63.2
mccabe==0.3.1
mock==1.0.1
pep8==1.5.7
py==1.4.29
pyflakes==0.8.1
pylint==1.4.3
pyserial==2.7
pytest==2.7.2
pytest-cov==1.8.1
six==1.9.0
wheel==0.24.0
